requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0

dash>=2.17.0
dash-bootstrap-components>=1.6.0
plotly>=5.22.0

pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.40

scikit-learn>=1.5.0
statsmodels>=0.14.0
shap>=0.45.0
lime>=0.2.0.1

wordcloud>=1.9.3
matplotlib>=3.9.0

requests>=2.32.0
feedparser>=6.0.11
python-dateutil>=2.9.0